+++ src/exceptions/CachePurgeFailed.php
@@ -1,6 +1,0 @@
-<?php
-
-namespace lenvanessen\cit\exceptions;
-
-class CachePurgeFailed extends \Exception
-{}